requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-core
python-dotenv
faiss-cpu
pypdf
